+++ client/src/pages/HostDetailPage/Followers/FollowersContainer.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import Axios from 'axios';
-import Loading from 'components/Loading/Loading';
+import Loading from 'components/Loading';
 import FollowersPresenter from './FollowersPresenter';
 import { FOLLOWER_SERVER } from 'utils/config';
